permisos para javaparser arreglado, además de quitar AST.java duplicado
