+++ modules/WorldCodecity/src/main/java/org/terasology/codecity/world/metrics/AST.java
@@ -31,15 +31,16 @@ public class AST extends VoidVisitorAdapter<Object> {
 	private int length;
 	private int[] linesLength;
 	private File file;
+	private CompilationUnit cu;
 
 	public AST(String location) {
 		super();
 		path = location;
 		methodCalls = new ArrayList<MethodCallExpr>();
 		fields = new ArrayList<FieldDeclaration>();
-		file = new File(path);
+		file = new File(location);
 		try {
-			CompilationUnit cu = JavaParser.parse(file);
+			cu = JavaParser.parse(file);
 			this.setGeneralMetrics(cu);
 			this.visit(cu, null);
 		} catch (ParseException | IOException e) {
@@ -58,10 +59,33 @@ public class AST extends VoidVisitorAdapter<Object> {
 		super.visit(n, arg);
 	}
 
+	/**
+	 * @return Representation of Code
+	 */
+	public CompilationUnit getCu() {
+		return cu;
+	}
+	
+	/**
+	 * @param text String to search
+	 * @return if .java contains indicated text
+	 */
+	public boolean contains(String text) {
+		return cu.toString().contains(text);
+	}
+
+	/**
+	 * @return Methods calls made in code
+	 */
 	public ArrayList<MethodCallExpr> getMethodCalls() {
 		return methodCalls;
 	}
 
+	/**
+	 * Sets general metrics for the class, such as
+	 * length, comments, packages, importations, etc
+	 * @param cu
+	 */
 	private void setGeneralMetrics(CompilationUnit cu) {
 		length = cu.getEndLine();
 		comments = cu.getComments();
@@ -83,6 +107,10 @@ public class AST extends VoidVisitorAdapter<Object> {
 		}
 	}
 
+	/**
+	 * Counts length for every line in code
+	 * @return Array with lengths
+	 */
 	private int[] countLineLength() {
 		BufferedReader br;
 		int[] resp = null;
@@ -102,26 +130,44 @@ public class AST extends VoidVisitorAdapter<Object> {
 		return resp;
 	}
 
+	/**
+	 * @return Length of code
+	 */
 	public int getLength() {
 		return length;
 	}
 
+	/**
+	 * @return ArrayList with comments included in code
+	 */
 	public List<Comment> getComments() {
 		return comments;
 	}
 
+	/**
+	 * @return ArrayList with imports made in code
+	 */
 	public List<ImportDeclaration> getImports() {
 		return imports;
 	}
 
+	/**
+	 * @return Package that contains code
+	 */
 	public PackageDeclaration getPackage() {
 		return pack;
 	}
 
+	/**
+	 * @return Variables/Fields of code
+	 */
 	public List<FieldDeclaration> getFields() {
 		return fields;
 	}
 
+	/**
+	 * @return Int[] with length of lines in the code
+	 */
 	public int[] getLinesLength() {
 		return linesLength;
 	}
